rename test type
